update gps code

+++ Project1/3년차/gps/gps_with_gyro2.cpp
@@ -346,7 +346,7 @@ double get_gpsData(char *chrBuf, int num)
         return lon;
     }
     else if (num==2){
-        lat = (tmp[7]<<24)|(tmp[6]<<16)|(tmp[5]<<8)|tmp[4];
+        lat = (double)((tmp[7]<<24)|(tmp[6]<<16)|(tmp[5]<<8)|tmp[4]);
         printf("* lat base : %lf\r\n\r\n", lat);
         return lat;
     }
